- switch to UTF-8; require all translations to use UTF-8 as well. - use UCS-4 instead of UCS-2 when editing/rendering text, since some   ancient languages are past the 16-bit range.

+++ trunk/engine/widget/keycodes.h
@@ -1,4 +1,4 @@
-/*	$Csoft: keycodes.h,v 1.10 2003/05/24 15:43:55 vedge Exp $	*/
+/*	$Csoft: keycodes.h,v 1.11 2003/06/15 08:54:19 vedge Exp $	*/
 /*	Public domain	*/
 
 #ifndef _AGAR_WIDGET_KEYCODES_H_
@@ -8,7 +8,7 @@
 struct keycode {
 	SDLKey	 key;
 	int	 modmask;
-	void	(*func)(struct textbox *, SDLKey, int, const char *, Uint16);
+	void	(*func)(struct textbox *, SDLKey, int, const char *, Uint32);
 	char	*arg;
 };
